서비스, 유스케이스 과세분화 command, query 구조로 리펙토링
