+++ legend-momo-city/src/main/java/com/wanted/momocity/auth/application/service/AuthQueryService.java
@@ -2,8 +2,12 @@ package com.wanted.momocity.auth.application.service;
 
 import com.wanted.momocity.auth.application.command.EmailVerifyCommand;
 import com.wanted.momocity.auth.application.port.EmailCodePort;
-import com.wanted.momocity.auth.application.usecase.EmailVerifyUsecase;
+import com.wanted.momocity.auth.application.port.LoadUserPort;
+import com.wanted.momocity.auth.application.usecase.AuthQueryUsecase;
 import com.wanted.momocity.auth.domain.exception.InvalidVerificationCodeException;
+import com.wanted.momocity.auth.domain.exception.UserNotFoundException;
+import com.wanted.momocity.auth.domain.model.User;
+import com.wanted.momocity.auth.presentation.api.response.LoginCompletedResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -13,10 +17,12 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 @Transactional
 @RequiredArgsConstructor
-public class EmailVerifyService implements EmailVerifyUsecase {
+public class AuthQueryService implements AuthQueryUsecase {
 
     private final EmailCodePort emailCodePort;
+    private final LoadUserPort loadUserPort;
 
+    // 이메일 인증
     @Override
     public void emailVerify(EmailVerifyCommand command) {
         String savedCode = emailCodePort.find(command.email());
@@ -36,5 +42,14 @@ public class EmailVerifyService implements EmailVerifyUsecase {
         emailCodePort.saveVerified(command.email(), 180L);  // 인증 성공하면 그 값을 3분동안 유지하고
         log.info("[email] 이메일 인증 완료 | email={}", command.email());
 
+    }
+
+
+    // 로그인 후 정보 전달
+    @Override
+    public LoginCompletedResponse getInfo(String userId) {
+        User loginUser = loadUserPort.findById(Long.parseLong(userId))
+                .orElseThrow(() -> new UserNotFoundException("사용자를 찾을 수 없습니다."));
+        return new LoginCompletedResponse(loginUser.getRole(),loginUser.getIsTempPwd(),loginUser.getNickname());
     }
 }
