--- legend-momo-city/src/main/java/com/wanted/momocity/auth/application/usecase/EmailVerifyUsecase.java
@@ -1,7 +1,0 @@
-package com.wanted.momocity.auth.application.usecase;
-
-import com.wanted.momocity.auth.application.command.EmailVerifyCommand;
-
-public interface EmailVerifyUsecase {
-    void emailVerify(EmailVerifyCommand command);
-}
