--- legend-momo-city/src/main/java/com/wanted/momocity/auth/application/usecase/GoogleLoginUsecase.java
@@ -1,9 +1,0 @@
-package com.wanted.momocity.auth.application.usecase;
-
-import com.wanted.momocity.auth.application.command.SocialLoginCommand;
-import com.wanted.momocity.auth.presentation.api.response.LoginResponse;
-
-public interface GoogleLoginUsecase {
-    LoginResponse socialLogin(SocialLoginCommand command);
-
-}
