--- legend-momo-city/src/main/java/com/wanted/momocity/auth/application/usecase/KakaoLoginUsecase.java
@@ -1,10 +1,0 @@
-package com.wanted.momocity.auth.application.usecase;
-
-import com.wanted.momocity.auth.application.command.SocialLoginCommand;
-import com.wanted.momocity.auth.presentation.api.response.LoginResponse;
-
-public interface KakaoLoginUsecase {
-
-    LoginResponse socialLogin(SocialLoginCommand command);
-
-}
