--- legend-momo-city/src/main/java/com/wanted/momocity/auth/application/usecase/LoginCompletedUsecase.java
@@ -1,7 +1,0 @@
-package com.wanted.momocity.auth.application.usecase;
-
-import com.wanted.momocity.auth.presentation.api.response.LoginCompletedResponse;
-
-public interface LoginCompletedUsecase {
-    LoginCompletedResponse getInfo(String email);
-}
