--- legend-momo-city/src/main/java/com/wanted/momocity/auth/application/usecase/LoginUsecase.java
@@ -1,8 +1,0 @@
-package com.wanted.momocity.auth.application.usecase;
-
-import com.wanted.momocity.auth.application.command.LoginCommand;
-import com.wanted.momocity.auth.presentation.api.response.LoginResponse;
-
-public interface LoginUsecase {
-    LoginResponse login(LoginCommand command);
-}
