--- legend-momo-city/src/main/java/com/wanted/momocity/auth/application/usecase/LogoutUsecase.java
@@ -1,8 +1,0 @@
-package com.wanted.momocity.auth.application.usecase;
-
-import com.wanted.momocity.auth.application.command.LogoutCommand;
-
-public interface LogoutUsecase {
-    void logout(LogoutCommand command);
-
-}
