--- legend-momo-city/src/main/java/com/wanted/momocity/auth/application/usecase/TempPasswordUsecase.java
@@ -1,9 +1,0 @@
-package com.wanted.momocity.auth.application.usecase;
-
-import com.wanted.momocity.auth.application.command.EmailSendCommand;
-
-public interface TempPasswordUsecase {
-
-    void sendTempPassword(EmailSendCommand command);
-
-}
