+++ legend-momo-city/src/test/java/com/wanted/momocity/auth/EmailVerifyTest.java
@@ -2,7 +2,7 @@ package com.wanted.momocity.auth;
 
 import com.wanted.momocity.auth.application.command.EmailVerifyCommand;
 import com.wanted.momocity.auth.application.port.EmailCodePort;
-import com.wanted.momocity.auth.application.service.EmailVerifyService;
+import com.wanted.momocity.auth.application.service.AuthQueryService;
 import com.wanted.momocity.auth.domain.exception.InvalidVerificationCodeException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -22,11 +22,11 @@ public class EmailVerifyTest {
     @Mock private EmailCodePort emailCodePort;
 
     @InjectMocks
-    private EmailVerifyService emailVerifyService;
+    private AuthQueryService authQueryService;
 
     @BeforeEach
     void setup() {
-        assertNotNull(emailVerifyService);
+        assertNotNull(authQueryService);
     }
 
     @Test
@@ -38,7 +38,7 @@ public class EmailVerifyTest {
         // when & then
         InvalidVerificationCodeException exception = assertThrows(
                 InvalidVerificationCodeException.class,
-                () -> emailVerifyService.emailVerify(new EmailVerifyCommand("test@test.com", "123456"))
+                () -> authQueryService.emailVerify(new EmailVerifyCommand("test@test.com", "123456"))
         );
 
         assertEquals("인증 코드가 만료되었습니다. 재발송 버튼을 눌러 인증코드를 다시 발급받아 입력해주세요.", exception.getMessage());
@@ -53,7 +53,7 @@ public class EmailVerifyTest {
         // when & then
         InvalidVerificationCodeException exception = assertThrows(
                 InvalidVerificationCodeException.class,
-                () -> emailVerifyService.emailVerify(new EmailVerifyCommand("test@test.com", "999999"))
+                () -> authQueryService.emailVerify(new EmailVerifyCommand("test@test.com", "999999"))
         );
 
         assertEquals("인증 코드가 올바르지 않습니다.", exception.getMessage());
